requirements.txt
openai>=1.12.0
anthropic>=0.18.0
httpx>=0.27.0
pyyaml>=6.0
rich>=13.7.0
click>=8.1.0
aiohttp>=3.9.0
tiktoken>=0.6.0
jinja2>=3.1.0
tabulate>=0.9.0
numpy>=1.26.0
